Add a test for linear interpolation of S parameter model
The commit adds a new test for linear interpolation of S-parameter
model. The S parameter model is represented by frequency domain
data in touchstone format. The data is generated from a RLC based
transmission line using S parameter analysis in Xyce. The frequencies
used in the test are different to the data in S parameter model and
interpolation is needed for the test. The linear interpolation is
tested. The gold standard matches the result from the original
transmission line for the circuit.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/46