Add a test for PSP DTA parameter on instance line
This is a test for this issue:

https://cee-gitlab.sandia.gov/Xyce/code/uur-proprietary/Xyce/-/issues/464
"Modify the PSP model so that DTA is an instance parameter"

DTA was elevated to an instance parameter in a patch that was
submitted.  This commit adds a test of that change.

Three netlists based on the "invert_psp" test are here.
  - invert_psp_testDTA0.cir runs exactly the invert_psp netlist with a
    `.options device temp=100` line added.  The model cards have DTA=0.
  - invert_psp_testDTA1.cir runs the same circuit but with
    `.options device temp=27` and DTA=73 in all the model cards.
  - invert_psp_testDTA2.cir runs the same circuit, but with DTA=73 on
    the mosfet instance lines and DTA=0 in the model cards, and with
    `.options device temp=27`.

All three netlists should end up running the simulation at T=100C.
Only the 0 and 1 versions would have worked before the patch I
submitted to the code.

The test runs all three, then compares the 1 and 2 versions to the 0
version.  It only passes if all three match.