Add a test for extrapolation methods in S parameter device
The commit adds a test for new extrapolation methods in S parameter
device. The S parameter data is generated from a RLC based
transmission line circuit. The S parameter analysis is used to
generate the data. The DEC sweep is used. And it does not have DC
data. Both lower and higher frequency points need extrapolation
and different methods are used in this circuit

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/701