Add more test for issue 701
The commit adds a circuit to test method 0 extrapolation in S parameter
device. The cut off method is not suitable in many cases.
But in this example, it is the best method to use at lower frequency
points, especially for short circuit current. The data is generated
from an EM simulator. The result with extrapolation is compared to
the circuit that has all the data and they match well

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/701