Set performance tag for circuits that will be used in testing
Remove the performance tag from some circuits that will not
be used in performance testing.  chip2.cir, ram2k.cir
memplus.cir and voter.cir will be used for performance
testing from this directory.  Also add a voter.cir.options
to specify a timelimt for that circuit as it takes
several hours

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/549