Scripts to setup and gather performance data
The script SetupXycePerformanceTests.py sets up a working
directory and writes the job queue submission scripts
needed for running the performance tests.

The script GatherXycePerformanceData.py traverses
the working directory from the setup script and
gathers all of the Xyce output metrics.  The data
is stored in a Python Pandas table and then exported
to a csv formatted file.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/549
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/547