Catch case where Xyce is killed on HPC due to time.
On the HPC systems, a job is killed when its alotted time
ends.  Xyce doesn't know when this occurs so it can't
output any special info in the normal output.  But looking
at the stats collected, the GaterXycePerformanceData script
can account for these tests.  When a job is killed do to
a time limit, the gather script will emit an error and not
add any of the data for that job to the data base.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/547
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/549