Test for recent .sens bugfix
This is a test for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/729

"Expression-based objective functions for .SENS conflate I(VA) with
V(VA) (nodes and devices w/ same names)"

This test consists of two netlists, which are both run and the results
compared.  They are identical, except that one of them has a node name
(VA) which matches a device name.   The other one does not.  Prior to
my fix, Xyce produced (incorrectly) different answers for these two
tests.  Now they (correctly) match.