Update performance testing scripts
Several updates to the setup script.  Fixed an issue where
the script had to be run from just above the Xyce_Regression
directory.  Made all of the user dependent queque submission
elements of the sbatch script options that can be set externally.
And the submission scripts echo the job name to the SLURM error
file so if the job fails due to time limit or out of memory, it
is easier to track that back to a given test.

On the gather metrics scipt, I added output data fields for the
machine on which the tests were run and the compiler used in
generating the code.  The gather script also catched Xyce run
time errors and will not add data from those run to any results
table.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/549