Add regression tests for BSIMCMG 111
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

These tests are identical to the tests for the 110, expcept that the
level/version number in the model cards has been changed to 111.

The gold standards are generated from Xyce, rather than another simulator
(which is what was done for the 110 and previous BSIM CMG models).
This is largely b/c the small set of benchmarks that was distributed
with the previous models (which included gold data) was no longer part
of the 111 distribution.

I suspect this is because the CMC test suite has superceded those old
benchmarks.  The Xyce implementation of the 111 passes all the CMC
tests, except for cases where the xyce-specific script itself is
busted.

As of this writing, the CMC test suite has 765 nmos tests and 765 pmos
tests.  The script problems for Xyce start at test 733, for both nmos
and pmos.  The test harness exits at that point.  I have not
exhaustively checked the remaining tests, but I've run a subset by hand
and they look OK.

So, I think Xyce passes most of the CMC tests, but it isn't practical to
put all of them into our nightly regression.

One thing we test, that isn't part of the CMC test suite is
sensitivities.  A couple of sensitivity tests for version 111
are included in this commit.

Types of tests that I have not included in this commit:

1. noise tests
2. M= multiplier tests
3. M= multiplier tests, but on subcircuits
4. other "MOD" choices, such as GEOMOD=5 (for GAA FETs)

I plan to create tests for those use cases, but it won't be today.