Add a noise test for CMG version 111
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

The tests I previously added lacked a noise test.  This adds one.

I patterned it off of the (as far as I could tell) the only other CMG
noise test in the test suite, which was for version 107.