Allow for +/-180 on phase output for FFT
Xyce uses the atan2() function to calculate the phase
angle for FFT analysis.  This function has a discontinuity
at +/-180 degrees which comes up in testing of DC signals.
A small numeric difference on the order 1e-16 can result
in the phase output being either -180 or +180.  These
are identical, but the file_compare.pl script didn't
treat them as such.  Now file_compare.pl allows
for the phases of +/-180 within abstol to be the same.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/476