Forgot this file yesterday
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

Yesterday I added a single noise test for the BSIM-CMG 111.2.1 device.
That test failed in nightly regression, because I forgot  to include the
model card.  This commit adds that model card.

Hopefully, this will be enough to fix the problem.