Add noise multiplicity test for BSIM-CMG 111.2.1
This commit is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/753
"Implement version 111.2.1 of the BSIM CMG device"

This is patterned after the noise  multiplier  (m=) tests in the
Xyce_Regression/Netlists/MULTIPLICITY_FACTOR/BSIM4_noise directory.
Two versions of the circuit are run.  One uses a single transistor with
m=6.  The other has six identical transistors in parallel.

Only the total ONOISE and INOISE are compared.