Update this file to add m=6
An older version of this file was committed yesterday, which
didn't include the m=6 multiplier.  In my local copy I had added
it, but it didn't get included in the main commit.  As a result,
this test was failing everywhere.  This should fix it.

This commit is for:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/729
"Expression-based objective functions for .SENS conflate I(VA) with
V(VA) (nodes and devices w/ same names)"