Update CSV noise output gold standard headers
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/613
"CSV output format should (maybe optionally) put quotes around variable/column names"

Noise output for CSV files has been fixed, so that the variable names
are now surronded by quotes.  This has been done to avoid conflating the
meaning of the comma separator, as many noise output variables include a
comma as part of their names.

This fix required that the first line of the gold standard for two noise
tests be updated.