Add a DC test for issue 721
The DC circuit tests the option gminscaling for b3soi device.
When the gmin scaling is turned off, the gmin used in b3soi
device is consistent to the gmin used in other nonlinear devices.
The result with default setting (gmin scaling on and gmin 1e-12)
is equivalent to the result with gmin scaling off and gmin 1e-18.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/721