Update test for new delimiters
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/613
"CSV output format should (maybe optionally) put quotes around variable/column names"

My recent code changes for this issue broke this test.  The reason was
that this test was intended to catch an incorrect delimiter, and the
"incorrect" delimiter was COLON, which is now supported.  So, the
expected error message was not generated.

I've updated the test so that the netlist now says DELIMITER=wrong,
which will trigger the expected error message.