Another test for AC sensitivity analysis
This is related to this issue:
"Expression-based objective functions for .SENS conflate I(VA) with
V(VA) (nodes and devices w/ same names)"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/729

It was discovered today that my previous fix for issue 729 created a new
bug in AC sensitivity analysis.  This wasn't caught because the AC
sensitivity tests weren't adequate.  This test corrects that issue.
Now, if the same mistake were to happen, this test would fail.

This commit also includes the beginnings of a test for issue 694, which
I had been working on before the above issue emerged.  That test isn't
done yet, but it is excluded so it shouldn't cause any harm for now.