Add test for using option scale w/FD sensitivities
This is related to this issue:
"Error in MOSFET numerical sensitivities, when using .option SCALE"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/694

This creates a simple test, based around the BSIM4.   It consists of 3
equivalent circuits.  They are:

1. bsim4_issue694.cir, which computes a BSIM4 circuit sensitivity using
   .SENS w.r.t. the device width, while using .option scale=1.0e-6

2. bsim4_issue694_ref.cir, which does the same as bsim4_issue694.cir,
   except that it doesn't use .option scale.

3. bsim4_issue694_FD.cir, which computes the sensitivity using a crude
   finite difference, by having two copies of the circuit in the
   netlist, with a slightly different parameter value.  So, this netlist
   doesn't invoke .SENS.

For netlists 1 and 2, the sensitivity option "forcedevicefd=1" is used.
This is because the whole point of the test is to ensure that
device-level finite differences (which are used when device-level
analytical sensitivies aren't available) work with .option scale.

Currently, the BSIM4 device doesn't have analytical sensitivities
implemented.  So, for the moment the forcedevicefd=1 command isn't
necessary.   However, this may change in the future, so to ensure that
this test continue to exercise numerical device sensitivies, this option
has been included.

The beginnings of this test were inadvertantly included in my previous
commit, that was primarily about a completely different test.

Also, note that I created the new tag "scale".  In addition to using
that tag on this new test for sensitivities, I went back and added it to
many of the previous .options scale related tests.