Fix the DEV_DOC and DEV_DOC_CAT test gold stds
My recent work on this issue:
"Implement DTEMP parameter for R, L, C, D, Q, M and J devices"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/343

Broke these tests.  It broke them in one way that was intentional, and
another way that was not intentional.  I just fixed the unintentional
issue in the source code (which had to do with the default value for
TEMP being "Ambient Temperature", which is set as 0.0 in the addPar
call).

This commit fixes the tests for the intentional breakage.  The parameter
DTEMP needed to be included in these tex file tables.