Updating this file, again.
This is part of my work on this issue:
"Implement DTEMP parameter for R, L, C, D, Q, M and J devices"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/343

Yesterday I updated this file to accomodate my changes for DTEMP.

However, it still didn't pass nightly regression.
Weirdly, when I do a "diff", while it says the DTEMP line is different, every
single characeter on the line looks the same (to me).  So, the difference
must be in a character not visible to the terminal.  Possibly it was a
whitespace difference.

Today I generated the file on my linux RHEL box.  Yesterday I generated
it on my OSX laptop.  This should fix it.