Add BJT DTEMP tests
This is part of my work on this issue:
"Implement DTEMP parameter for R, L, C, D, Q, M and J devices"
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/343

In my last commit, I claimed it would be the last set of tests on this
issue.  I was wrong.  I still had not created level=1
BJT tests.   This commit adds them.