Create two tests on nonlinear VCCS multipliers
These tests are necessary but not sufficient.  More tests coming.

This is to test this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/831
"Add multiplier (M=) to the nonlinear VCCS device (ie, the Bsrc) to support
implicit multilpiers"

Also included in this commit is the addition of the "vccs" tag to
various other pre-existing tests.   As my recent changes are to
different versions of the VCCS device, it is useful to test a bunch of
them with a single tag.