Add nonlinear VCCS implicit multiplier test
This is in support of:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/831
"Add multiplier (M=) to the nonlinear VCCS device (ie, the Bsrc) to support
implicit multilpiers"

Earlier today I added two tests for this feature.   However, neither
test included an implicit subcircuit multiplier.  The first test
(vccs_nl1_m.cir) didn't use a subcircuit.  The second test
(vccs_nl2_m.cir) used a subcircuit, but all instances of the "M="
multiplier were explicitly set in the netlist.

In this new test, the M= parameter on the G-source is not specified.  It
is assumed to be implicitly assigned the one from the subcircuit parameter
list.

In other words, compared to vccs_nl2_m.cir, I changed this:

Xtest 3 0 2 0 gsrcSub m=10

.subckt gsrcSub A B C D M=1
G A B cur='v(C,D)*0.02' M='M'
.ends

to this:

Xtest 3 0 2 0 gsrcSub m=10

.subckt gsrcSub A B C D M=1
G A B cur='v(C,D)*0.02'
.ends

A few minutes ago I pushed code changes that allow this second use case
to work, so now it is safe to add the test.