Add another implicit multiplier test for nl VCCS
This is to fix this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/831
"Add multiplier (M=) to the nonlinear VCCS device (ie, the Bsrc) to support
implicit multilpiers"

This new (and probably last for now) test uses implicit subcircuit
multipliers in nested subcircuits.  Similar to the previous tests, the
implicit nested multiplier works out to the equivalent of m=10.