Test for Bsrc multipliers (implicit and otherwise)
This is to test this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/831
"Add multiplier (M=) to the nonlinear VCCS device (ie, the Bsrc) to support
implicit multilpiers"

Earlier today I added a number of tests for multipliers in the nonlinear
VCCS (G-source) device.  However, this wasn't sufficient testing,
because the nonlinear VCCS device gets transformed into a Bsrc
internally.  So, to support multipliers in the VCCS device, it made
sense to also support them in the Bsrc device, at least for the "current
source" version of that device.

This commit adds 4 tests that are the equivalent of what I added to the
VCCS directory earlier today.  They test both implicit and explicit
multipliers on the Bsrc, in a few different configurations.

In terms of code coverage the main difference between these new Bsrc
tests and the VCCS tests is in the parser.  In that respect, the VCCS is
more fragile, as the parser code that transforms it to a Bsrc is
fragile.