Update these tests to make bug fix effect more obvious
The hysteresis added to the switch device was difficult
to see in the original versions of these test cases.
I've limited the test output to just the control signal
and the current through the switch so the cause and
effect are the only outputs.  Also, the hystereis is
active over about 90% of the switching range so the
effect is more obvious.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/752