Fix tests with conflicting files
These tests were either using the wrong input files through a
typo or were sharing input files with another test.  When tests
were run in parallel these tests could fail as another test
would overwrite some simulation output.  In the case of the
EMBEDDED and SAMPLING directories it was a clear case where
a test template was used but the correct circuit wasn't set
(i.e. it was global rather then normal or vice versa).  In
the case of but456output.cir.sh this test used a base case
that was also used by another test.  So that base case was
duplicatd and given a unique name.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/658

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/658