Add netlist file to exclude list
In an earlier commit I added a test file that is used by a
shell script test.  But that added netlist doesn't have the
same name as the shell script, so the "run_xyce_regression"
script assumed it was a new test that is run directly in Xyce.
This caused an error on many platforms.  Added it to the exclude
file now.

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/658