Remove excluded netlist as a test.
The netlist converter_baseline_output.cir is called from a shell
script test of a different name.  Since it was accidently left
out of the exclude file, it was added as a test.  It should not
be an independent circuit test so it was removed from the
CMakeLists.txt file

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/658