Fix duplicate file use
The test tran-prn-noheaderfooter.cir.sh was using a test file,
tran-prn.cir, from another test as a gold standard.  When the
tests are run in parallel this use of another test's input files
causes random errors as two copies of Xyce are reading and
writing output to files named tran-prn.cir and tran-prn.cir.prn.
Fixed by giving the test a unique input file to use for its
gold standard.