Add announcement for open source users
After the recent reset of the master and develop branches to an orphaned
commit, developers will likely receive an ‘unrelated histories’ error
message from git when updating their local clones of this repo. This
commit includes a notice about the branch change and how to address the
error in the README.md file. Since the repo’s README.md file is rendered
on the main repository page in GitLab and GitHub, that should help users
find the message.