Fix spurious diff error
This test uses diff to compare the output of two HB runs that
should be identical.  The output is identical in most cases
but I find with Apple's native FFT calls that there can
be two values that differ by 1e-8.  This really looks like
some optimization that effects the last bit in the outout and
is not significant.  Using the precision option on the print
line fixes the issue.