Minor fixes related to doping interpolation
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/884

I recently fixed how doping interpolations are done in Xyce for one
dimensional PDE devices.

Both of these tests use this feature and were potentially affected by
this change.  In particular, the npn_bjt_pde.cir test started to fail,
because it didn't converge on the very last DC sweep point.  This was
easy to fix, by slightly lowering the "chop" value in the netlist.

I also noticed that both of these circuits had the "tecplotlevel"
parameter set to 1, and this wasn't necessary.  It resulted in a large
number of tecplot files being output during the test, but they were not
part of the test.  So, I have changed the "tecplotlevel" variable to
zero in both netlists.