Fix tests for multi config generators.
Multi-config build generators place the build in
a subdirectory called "Release" or "Debug" or
other names.  These tests look for extra build
programs by parsing the Xyce full path but
get things wrong when a multi-config build
is used.  The best approach might be to make
new CMakeLists.txt scripts for these but for
now a small change makes them work with a
ninja Release target

https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/453