Test for bugfix in rawfiles
This is in support of this issue;
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/888

The issue was that if a netlist contained transient adjoint
sensitivity commands, and the user chose to use the command line option
for rawfiles ("-r"), the resulting rawfile got mangled.

It got mangled, in that Xyce would output the standard transient
rawfile, for the non-adjoint part of the calculation.  But it
would then tack on to the end of the file some adjoint stuff in
the STD (i.e. *prn) format.  So, you'd have a binary raw file with a
*prn file at the end.

This was happening in part b/c raw file output is not yet supported for
transient adjoints.  The adjoint output it *did* put there was all
zeros.   In the way that adjoints work, they are purely a post-process
step, and the output needs to be done separately.

Anyway, this was fixed a while ago and now this test ensures that
it is fixed.