End running command line test in parallel.  Update ctest files.
The tests in the directory CommandLine test error conditions
and command line option output.  But when run in parelle with
may other paralel jobs running they often fail due to MPI
stalls.  This test doesn't test any parallel capabilities in
Xyce, so this commit removes the parallel tag for this test.

This commit also updates the CMakeLists files for the tests
to better consolidate the test set-up, file copy operations.
This makes the files more human readible and speeds up
test setup.