Move the tests for BSIM SOI v461 to new directory
A new version of the BSIM SOI has been added to Xyce.

That new version is 4.7.0.

This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/958
"Implement the BSIM SOI v4.7.0 in Xyce"

As a result, the version 4.7.0 will now tested in the B4SOI directory,
and the 4.6.1 will be tested in the B4SOI_461 directory.

This commit copies the 4.6.1 tests to to B4SOI_461 directory, and
updates the level numbers from 70 to 70461.