Revert level numbers; BSIM SOI v4.6.1 and v4.7.0
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/958
"Implement the BSIM SOI v4.7.0 in Xyce"

The new version 4.7.0 does not produce similar enough results to the
version 4.6.1. on some of our internal tests, for it to be made the
default LEVEL=70.  This may change later, but it needs more evaluation
before we can make the change.

Also, our 4.7.0 implementation has a problem with SHMOD=1.  A lot of
Jacobian derivative terms are wrong when that model option is used.
This prevents the invert_b4soi.cir circuit from running properly.

So, for now I have changed the source code and tests to reset
LEVEL=70 to point at version 4.6.1.   The v4.7.0 will use
LEVEL=70470 for now.