Commenting out the invert_b4soi test
My intent was for this test to be (hopefully temporarily)
excluded, so I put it in the "exclude" file.

That is sufficient to exclude it when using the old test
framework.  However, ctest doesn't use that file.  To
exclude a circuit from the ctest framework, the commands
in the CMakeLists.txt file must be modified.

This commit comments out the commands related to
the invert_b4soi test.  I chose not to delete it, because
I have hope that this test can be restored after some
debugging.