Test plugin feature with more than two plugins
The plugin feature did not work when "-plugin" was given a comma
separated list that contained more than two library paths.

This test is an exact duplicate of DiodeClipper, but instead of
building all the toy verilog models into a single shared library, it
builds each one into its own, then gives "-plugin" a list of four
libraries.

Prior to the fix of [Xyce issue 148](https://github.com/Xyce/Xyce/issues/148)
this fails.

I have ONLY created a test that works with run_xyce_regression on
installed binaries using buildxyceplugin, because I could not figure
out how to make the right mods to CMakeLists.txt *and* the ctest
infrastructure to support building more than one plugin library and
accumulating the list into a single argument for "-plugin".  I leave
that as an exercise for the Xyce team.

[Xyce issue 148](https://github.com/Xyce/Xyce/issues/148)