Fix a typo in test script
This is for this issue:
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/888
"Xyce raw files (if specified via the command line "-r" argument) can
incorrectly include adjoint sentivity text"

This issue was fixed a while ago in the Xyce source code.  However, the
test that I created for it had a mistake in it.  The mistake was
comparing a file to itself, rather than to the reference result.
This commit fixes the issue, and fortunately, the test still passes.