Add several analytical sensitivity Vsrc tests
This commit is related to this issue
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/849
"Sensitivities don't work with PWL source"

I recently implemented analytical device sensitivity functions in the
independent source devices (Vsrc and Isrc).  Previously they were not
implemented, so they either had to be computing using device-level
finite difference derivatives, or with circuit-wide finite difference
derivatives.  For some sources (PWL and PAT) none of these options
worked for various reasons.

There was an existing test (vsrcSensTest.cir) which tested an
independent source using device-level finite difference derivatives.
For the first test of analytical derivatives, I chose to use this
test, but to force it to use analytical derivatives.  This is the
first new test.  This test is nice because it has an analytical
solution.

This commit adds the following tests: one for the PULSE, one for PWL,
one for SFFM and one for SIN Vsrc devices.  These are completely new,
not derived from previous tests.  All of them compare analytical-based
sensitivities to finite-difference (internal) based sensitivities.

This commit also updates the CMakeLists.txt for the new tests.