Disabling this (now obsolete) test
This is related to this issue.
https://cee-gitlab.sandia.gov/Xyce/backlog/project-backlog/-/issues/990
"Usage text misrepresents how the "-param" feature can be used"

An external user pointed out that when "Xyce -h" is run, the behavior it
describes for Xyce -param is not correct.   It claims that the user can
optionally have device and level arguments after the "-param" argument,
and then Xyce will print params for ONLY that device.

However, part of why it didn't work is the behavior tested by this
regression test.  Many years ago (2006) a user wanted Xyce -param to
print the full list of parameters, even if they specified a netlist on
the command line.

Reading thru the original bug, it is clear the user really just wanted
the equivalent of Xyce -syntax and/or Xyce -norun.  They wanted
diagnostic information about the specified circuit, including
device count.  So, I think that fixing Xyce -param in this way
was a mistake.

I have modified Xyce to behave in the manner described by Xyce -h.
That breaks this old regression test, so I am disabling it, as this
is an intended change.