Allow tests to run just Xyce to determin pass/fail
Add in a testing pathway so that one can just run Xyce on
a circuit and check the Xyce output for pass/fail.
run_xyce_regression would previously require that either
at test is run from a shell script or it had a gold standard
in the output data directory.  But some tests like lead
current tests have very simple gold standards that can be
entirely specified in a Xyce netlist.  So this commit enables
that testing methodology

https://cee-gitlab.sandia.gov/Xyce/backlog/bugs/-/issues/73