Try multiple runs of this test if sampled time is too long
This test should run quickly on a typical machine but on a
heavily subscribed machine the test may take longer than the
test limited time of 10 seconds.  In the event of failure, try
running the test again to better assess if the test has really
failed or if the sytem load was just temporally too high.