Set up top level CMakeLists.txt file for installed testing
Set up the top level CMakeLists.txt file so that regression
testing can be done on an installed copy of Xyce.  Primary
changes are using the cache variable XYCE_BINARY in place of
the generator expression $<TARGET_FILE:Xyce> at the target
will not exist if Xyce isn't built.  Second, deduce the
value of build flags like Xyce_PARALLEL_MPI from the output
of Xyce -capabilities so that tests are run properly.